Distribution.Client.SetupHooks.CallHooksExe

Documentation

callHooksExe

Arguments

:: forall inputs outputs. HookIO inputs outputs

=> FilePath
path to hooks executable

-> String
name of the hook to run

-> inputs
argument to the hook

-> IO outputs

Call an external hooks executable in order to execute a Cabal Setup hook.

externalSetupHooks :: FilePath -> SetupHooks

Construct a SetupHooks that runs the hooks of the external hooks executable at the given path through the CLI.

This should only be used at the final step of compiling a package, when we have all the hooks in hand. The SetupHooks that are returned by this function cannot be combined with any other SetupHooks; they must directly be used to build the package.

externalSetupHooksABI :: FilePath -> IO HooksVersion

The ABI of an external hooks executable.

This information is used to handshake before further communication, in order to avoid a cascade of errors with mismatched Binary instances.

buildTypeSetupHooks :: Maybe (SymbolicPath CWD (Dir Pkg)) -> SymbolicPath Pkg (Dir Dist) -> BuildType -> SetupHooks

The SetupHooks associated to a particular BuildType.

*Warning:** for build-type: Hooks, this does not include the pre-build hooks. Those can be retrieved with buildTypePreBuildHooks.

buildTypePreBuildHooks :: Maybe (SymbolicPath CWD (Dir Pkg)) -> SymbolicPath Pkg (Dir Dist) -> BuildType -> PreBuildComponentInputs -> IO [MonitorFilePath]

The pre-build hooks obtained by communication with an external hooks executable.

runExternalPreBuildRules :: FilePath -> PreBuildComponentInputs -> IO [MonitorFilePath]

Run all pre-build rules coming from an external hooks executable at the given filepath.

TODO: in the future, we will want to keep track of the dependency graph ourselves, and when re-building, only re-build what we need (instead of re-running all rules).

hooksProgFilePath :: Maybe (SymbolicPath CWD (Dir Pkg)) -> SymbolicPath Pkg (Dir Dist) -> FilePath

The path to the external hooks executable.