Distribution.Client.SetupHooks.HooksExe

Documentation

hooksMain :: SetupHooks -> IO ()

Create a hooks executable given SetupHooks:

the first two argument are references to input & output communication handles,

the second argument is the hook type.

The hook reads binary data passed to it over the input handle, decodes it, runs the hook, and encodes its result to binary, writing the result to the output handle.